requirements.txt
selenium
edge-tts
moviepy
requests
webdriver-manager
msedge-selenium-tools